updt

+++ src/main/java/com/grapplesoft/meil_backend/models/request/StatusRequest.java
@@ -3,7 +3,7 @@ package com.grapplesoft.meil_backend.models.request;
 public record StatusRequest(
         Integer statusid,
         String status,
-        Integer createuserid,
+        Long createuserid,
         Integer edituserid,
         boolean isdeleted
 ) {
